Switch DataObjectRoute's fields from int32 to int64 (and the same fields in other types)

+++ route4me-csharp-sdk/Route4MeSDKLibrary/DataTypes/MemberResponse.cs
@@ -36,7 +36,7 @@ namespace Route4MeSDK.DataTypes
         ///     Member ID
         /// </summary>
         [DataMember(Name = "member_id")]
-        public int? MemberId { get; set; }
+        public long? MemberId { get; set; }
 
         /// <summary>
         ///     User's API key
@@ -60,7 +60,7 @@ namespace Route4MeSDK.DataTypes
         ///     Geofence polygon size
         /// </summary>
         [DataMember(Name = "geofence_polygon_size")]
-        public int? GeofencePolygonSize { get; set; }
+        public long? GeofencePolygonSize { get; set; }
 
         /// <summary>
         ///     Geofence onsite trigger time (seconds)
